Se corrige consulta de estado de envio y dte

+++ TestUnitarios/SII.cs
@@ -7,15 +7,15 @@ namespace TestUnitarios
     public class SII
     {
         Handler handler = new Handler();
-        private string pathCertificadoValido = System.IO.Path.Combine("Files","CertificadoGonzalo2021.pfx");
-        private string pathCertificadoVencido = System.IO.Path.Combine("Files","CertificadoTest.pfx");
+        private string pathCertificado = System.IO.Path.Combine("Files","CertificadoDigital.pfx");
         private string pathEnvioDTE = System.IO.Path.Combine("Files","ENVIO_DTE_REST.xml");
         private string pathToken = System.IO.Path.Combine("Files", "tkn.dat");
 
         [Fact]
         public async System.Threading.Tasks.Task ObtenerTokenAsync()
         {
-            string token = await SimpleAPI.WS.Autorizacion.Autenticar.GetTokenAsync(pathCertificadoValido, Ambiente.AmbienteEnum.Produccion, pathToken, "Pollito702");
+            if (!System.IO.File.Exists(pathCertificado)) throw new Exception("No existe certificado digital");
+            string token = await SimpleAPI.WS.Autorizacion.Autenticar.GetTokenAsync(pathCertificado, Ambiente.AmbienteEnum.Produccion, pathToken, "Pollito702");
             System.IO.File.Delete(pathToken);
 
             Assert.True(!string.IsNullOrEmpty(token));
@@ -24,18 +24,20 @@ namespace TestUnitarios
         [Fact]
         public async System.Threading.Tasks.Task ObtenerEstadoDTEAsync()
         {
+            if (!System.IO.File.Exists(pathCertificado)) throw new Exception("No existe certificado digital");
             var ambiente = Ambiente.AmbienteEnum.Produccion;
             var entity = new SimpleAPI.SII.GetEstadoEntity("17096073-4", "76269769-6", "3671414-K", new DateTime(2021, 5, 4), 33, 85, 75225);
-            var estadoDTE = await SimpleAPI.WS.Estado.EstadoDTE.GetEstadoDTEAsync(entity, pathCertificadoValido, ambiente, pathToken, "Pollito702");
+            var estadoDTE = await SimpleAPI.WS.Estado.EstadoDTE.GetEstadoDTEAsync(entity, pathCertificado, ambiente, pathToken, "Pollito702");
             Assert.True(estadoDTE.Ok);
         }
 
         [Fact]
         public async System.Threading.Tasks.Task ObtenerEstadoEnvioAsync()
         {
+            if (!System.IO.File.Exists(pathCertificado)) throw new Exception("No existe certificado digital");
             var ambiente = Ambiente.AmbienteEnum.Produccion;
             var entity = new SimpleAPI.SII.GetEstadoEnvioEntity("76269769-6", "4942604664");
-            var estadoDTE = await SimpleAPI.WS.Estado.EstadoEnvio.GetEstadoEnvioAsync(entity, ambiente, pathToken, pathCertificadoValido, "Pollito702");
+            var estadoDTE = await SimpleAPI.WS.Estado.EstadoEnvio.GetEstadoEnvioAsync(entity, ambiente, pathToken, pathCertificado, "Pollito702");
 
             Assert.True(estadoDTE.Ok);
         }
@@ -43,8 +45,9 @@ namespace TestUnitarios
         [Fact]
         public async System.Threading.Tasks.Task EnviarAsync()
         {
+            if (!System.IO.File.Exists(pathCertificado)) throw new Exception("No existe certificado digital");
             var ambiente = Ambiente.AmbienteEnum.Certificacion;
-            var envioDTE = await SimpleAPI.WS.Envio.EnvioDTE.EnviarAsync("17096073-4", "76269769-6", pathEnvioDTE, pathCertificadoValido, ambiente, pathToken, "Pollito702");
+            var envioDTE = await SimpleAPI.WS.Envio.EnvioDTE.EnviarAsync("17096073-4", "76269769-6", pathEnvioDTE, pathCertificado, ambiente, pathToken, "Pollito702");
 
             Assert.True(envioDTE.Ok);
         }
